+++ Scripts/Examples/RL-DDM.py
@@ -11,6 +11,7 @@ import random
 
 # CONSTRUCTION:
 import psyneulink.core.components.functions.integratorfunctions
+import psyneulink.core.components.functions.learningfunctions
 
 input_layer = pnl.TransferMechanism(
     size=2,
@@ -47,7 +48,8 @@ p = pnl.Process(
     default_variable=[0, 0],
     # pathway=[input_layer, np.array([[1],[-1]]), action_selection],
     pathway=[input_layer, pnl.IDENTITY_MATRIX, action_selection],
-    learning=pnl.LearningProjection(learning_function=pnl.Reinforcement(learning_rate=0.5)),
+    learning=pnl.LearningProjection(learning_function=psyneulink.core.components.functions.learningfunctions
+                                    .Reinforcement(learning_rate=0.5)),
     target=0
 )
