+++ Scripts/Models/MontagueDayanSejnowski96.py
@@ -10,6 +10,7 @@ import psyneulink as pnl
 # needed for 3d plotting in full_experiment and response_extinction
 from mpl_toolkits import mplot3d
 
+import psyneulink.core.components.functions.learningfunctions
 import psyneulink.core.components.functions.transferfunctions
 
 
@@ -54,7 +55,7 @@ def model_training():
         matrix=np.full((60, 60), 0.0)
     )
     learning_projection = pnl.LearningProjection(
-        learning_function=pnl.TDLearning(learning_rate=0.3)
+        learning_function=psyneulink.core.components.functions.learningfunctions.TDLearning(learning_rate=0.3)
     )
 
     p = pnl.Process(
@@ -153,7 +154,7 @@ def model_training_full_experiment():
     )
 
     learning_projection = pnl.LearningProjection(
-        learning_function=pnl.TDLearning(learning_rate=0.3)
+        learning_function=psyneulink.core.components.functions.learningfunctions.TDLearning(learning_rate=0.3)
     )
 
     p = pnl.Process(
@@ -246,7 +247,7 @@ def model_training_response_extinction():
     )
 
     learning_projection = pnl.LearningProjection(
-        learning_function=pnl.TDLearning(learning_rate=0.3)
+        learning_function=psyneulink.core.components.functions.learningfunctions.TDLearning(learning_rate=0.3)
     )
 
     p = pnl.Process(
