+++ psyneulink/library/components/mechanisms/adaptive/control/lvoccontrolmechanism.py
@@ -240,7 +240,8 @@ from enum import Enum
 
 import numpy as np
 
-from psyneulink.core.components.functions.function import BayesGLM, ModulationParam, _is_modulation_param, is_function_type
+from psyneulink.core.components.functions.function import ModulationParam, _is_modulation_param, is_function_type
+from psyneulink.core.components.functions.learningfunctions import BayesGLM
 from psyneulink.core.components.functions.optimizationfunctions import OBJECTIVE_FUNCTION, SEARCH_SPACE, \
     GradientOptimization
 from psyneulink.core.components.mechanisms.adaptive.control.controlmechanism import ControlMechanism
