+++ psyneulink/library/components/mechanisms/processing/transfer/lcamechanism.py
@@ -139,10 +139,9 @@ from collections import Iterable
 import numpy as np
 import typecheck as tc
 
-from psyneulink.core.components.functions.function import max_vs_avg, max_vs_next
+from psyneulink.core.components.functions.selectionfunctions import max_vs_avg, max_vs_next, MAX_VS_NEXT, MAX_VS_AVG
 from psyneulink.core.components.functions.integratorfunctions import LCAIntegrator
 from psyneulink.core.components.functions.transferfunctions import Logistic
-from psyneulink.core.components.functions.selectionfunctions import SelectionFunction
 from psyneulink.core.components.mechanisms.processing.transfermechanism import _integrator_mode_setter
 from psyneulink.core.components.states.outputstate import PRIMARY, StandardOutputStates
 from psyneulink.core.globals.keywords import BETA, ENERGY, ENTROPY, FUNCTION, INITIALIZER, LCA_MECHANISM, NAME, NOISE, OUTPUT_MEAN, OUTPUT_MEDIAN, OUTPUT_STD_DEV, OUTPUT_VARIANCE, RATE, RESULT, TIME_STEP_SIZE
@@ -150,9 +149,7 @@ from psyneulink.core.globals.parameters import Param
 from psyneulink.core.globals.preferences.componentpreferenceset import is_pref_set
 from psyneulink.library.components.mechanisms.processing.transfer.recurrenttransfermechanism import RecurrentTransferMechanism
 
-__all__ = [
-    'LCAMechanism', 'LCAMechanism_OUTPUT', 'LCAError', 'MAX_VS_AVG', 'MAX_VS_NEXT',
-]
+__all__ = ['LCAMechanism', 'LCAMechanism_OUTPUT', 'LCAError']
 
 
 class LCAError(Exception):
@@ -161,9 +158,6 @@ class LCAError(Exception):
 
     def __str__(self):
         return repr(self.error_value)
-
-MAX_VS_NEXT = 'max_vs_next'
-MAX_VS_AVG = 'max_vs_avg'
 
 # This is a convenience class that provides list of standard_output_state names in IDE
 class LCAMechanism_OUTPUT():
