+++ tests/learning/test_reinforcement.py
@@ -3,7 +3,7 @@ import numpy as np
 import pytest
 
 from psyneulink.core.components.functions.function import PROB
-from psyneulink.core.components.functions.function import Reinforcement
+from psyneulink.core.components.functions.learningfunctions import Reinforcement
 from psyneulink.core.components.functions.transferfunctions import SoftMax
 from psyneulink.core.components.mechanisms.processing.transfermechanism import TransferMechanism
 from psyneulink.core.components.process import Process
